requirements.txt
crewai>=1.6.1
fastapi>=0.110.0
langgraph>=0.1.11
litellm>=1.80.8
pdfplumber>=0.11.4
pydantic-settings>=2.2.1
email-validator>=2.1.0
python-multipart>=0.0.9
uvicorn>=0.23.2
sqlalchemy>=2.0.29
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0openai>=1.3.0